remove id function
