+++ x/identity/keeper/msg_server_id.go
@@ -8,14 +8,32 @@ import (
 	errorsmod "cosmossdk.io/errors"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+
+	"crypto/rand"
+	"math/big"
 )
+
+func generateSecureRandomID(length int) (string, error) {
+	const charset = "abcdefghijklmnopqrstuvwxyz" +
+		"ABCDEFGHIJKLMNOPQRSTUVWXYZ" +
+		"0123456789"
+	b := make([]byte, length)
+	for i := range b {
+		num, err := rand.Int(rand.Reader, big.NewInt(int64(len(charset))))
+		if err != nil {
+			return "", err
+		}
+		b[i] = charset[num.Int64()]
+	}
+	return "did:sovid:" + string(b), nil
+}
 
 func (k msgServer) CreateId(goCtx context.Context, msg *types.MsgCreateId) (*types.MsgCreateIdResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
 
 	// Check if the value already exists
 
-	newdid, diderr := k.CreateNewDid()
+	newdid, diderr := generateSecureRandomID(40)
 
 	if diderr != nil {
 		return nil, errorsmod.Wrap(diderr, "did error, please try again")
